Precinct scanner API
- Adds a new separate express application for precinct scanner endpoints to the scan service
- Moves the existing application to its own central scanner file
